requirements.txt
flask
transformers
torch
scikit-learn
pdfplumber
google-genai
spacy